refactor(audio-reader): fold collection modals into this domain
